backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
google-genai==1.9.0
python-multipart==0.0.20
httpx==0.28.1
pdfplumber==0.11.4
beautifulsoup4==4.12.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
numpy==2.2.1
fpdf2==2.8.3
